+++ galleria/serializer/exceptions.py
@@ -1,0 +1,9 @@
+"""Exceptions for photo collection serialization."""
+
+
+class ManifestValidationError(Exception):
+    pass
+
+
+class ManifestNotFoundError(Exception):
+    pass